Changed counter type in db

+++ db/migrations/000001_create_metrics_table.up.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS  metrics(
     id serial PRIMARY KEY,
     metric_id VARCHAR(256) NOT NULL UNIQUE,
     metric_type VARCHAR(50),
-    metric_delta INTEGER,
+    metric_delta BIGINT,
     metric_value DOUBLE PRECISION,
     hash VARCHAR(300)
 );
